refactor(data): split regiment economy catalog by era (#4292)
Slice B — move 29 regiment rows into era libraries behind the stable
RegimentEconomyCatalog façade. Add frozen catalog digest test for AC3
semantic parity.
